pyproject.toml
[tool.poetry]
name = "dblinter"
version = "0.0.0"
description = "A static analysis tool for postgresql databases"
authors = [
    "Stephane Defenin <stephane.defenin@gmail.com>",
    "Pierre-Marie Petit",
    "Grégoire Waymel",
    "DBRE awesome contributors"
    ]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/decathlon/dblinter"
classifiers = [
    'Topic :: Utilities',
    'Topic :: Software Development :: Quality Assurance',
    'Topic :: Software Development :: Testing',
    'Topic :: Software Development :: Build Tools',
    'Programming Language :: Python :: 3.12',
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: OS Independent',
    'Intended Audience :: Developers',
    ]
keywords = ["postgresql", "static analysis", "sonar", "sarif", "code quality", "sql", "database"]

[tool.poetry.dependencies]
python = "^3.12"
rich = "^12.6.0"
typer = "*"
psycopg2-binary = "^2.9.5"
google-cloud-storage = "^2.7.0"
pyyaml = "^6.0"
sarif-om = "^1.0.4"
# sqlalchemmy and greenlet are required by testcontainers-python for some exceptions
# https://testcontainers-python.readthedocs.io/en/latest/core/README.html
sqlalchemy = "^1.4.46"
greenlet = "^3.0.3"
types-pyyaml = "^6.0.12.3"
jschema-to-python = "^1.2.3"
pydantic = "^2.8.2"
pydantic-yaml = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.2"
black = "^22.10"
isort = "^5.10"
pytest = "^7.2"
pytest-mock = "^3.10"
freezegun = "^1.2"
pep8-naming = "^0.13"
testcontainers = "^3.7.1"
pytest-cov = "^4.0.0"
pre-commit = "^3.0.0"
ruff = "^0.0.284"


[tool.poetry.scripts]
dblinter = "dblinter.__main__:cli"

[tool.pytest.ini_options]
addopts = ""
filterwarnings = [
# ignore deprecation warnings from setuptools required by typer
"ignore:Deprecated call to `pkg_resources\\.declare_namespace\\('.*'\\):DeprecationWarning",
"ignore:pkg_resources is deprecated as an API\\.\\('.*'\\):DeprecationWarning",
]

[tool.isort]
profile = "black"

[tool.ruff]
ignore = ["B008", "C901", "E203", "E266", "E501"]
select = ["B","C","E","F","W"]

[tool.pylint]
max-line-length=120

[tool.pylint.'MESSAGES CONTROL']
disable=[
    "line-too-long",
    "too-many-arguments",
    "too-many-locals",
    "too-many-branches",
    "attribute-defined-outside-init",
    "invalid-name",
    "missing-class-docstring",
    "missing-module-docstring",
    "missing-function-docstring",
    "duplicate-code",
    "import-error",
    # sarif_om generate unsubscriptable-objecte error. This is fixed by this pull request:
    # https://github.com/microsoft/sarif-python-om/pull/6/ not yet released
    "unsubscriptable-object"
]
notes=[
    "fixme",
    "missing-function-docstring"
]
extension-pkg-whitelist = "pydantic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry-core.masonry.api"

